ref(perf): Transaction summary chart widget query hook
