+++ static/app/views/performance/transactionSummary/transactionOverview/eapChartsWidget.tsx
@@ -3,19 +3,11 @@ import styled from '@emotion/styled';
 
 import {CompactSelect} from 'sentry/components/core/compactSelect';
 import {t} from 'sentry/locale';
-import type {DataUnit} from 'sentry/utils/discover/fields';
-import {DiscoverDatasets} from 'sentry/utils/discover/types';
-import {decodeScalar} from 'sentry/utils/queryString';
-import {MutableSearch} from 'sentry/utils/tokenizeSearch';
-import {useLocation} from 'sentry/utils/useLocation';
-import type {TimeSeries} from 'sentry/views/dashboards/widgets/common/types';
-import {Area} from 'sentry/views/dashboards/widgets/timeSeriesWidget/plottables/area';
 import {TimeSeriesWidgetVisualization} from 'sentry/views/dashboards/widgets/timeSeriesWidget/timeSeriesWidgetVisualization';
 import {Widget} from 'sentry/views/dashboards/widgets/widget/widget';
-import {useSpanIndexedSeries} from 'sentry/views/insights/common/queries/useDiscoverSeries';
-import {SpanIndexedField} from 'sentry/views/insights/types';
+import {useWidgetChartQuery} from 'sentry/views/performance/transactionSummary/transactionOverview/useWidgetChartQuery';
 
-enum EAPWidgetType {
+export enum EAPWidgetType {
   DURATION_BREAKDOWN = 'duration_breakdown',
   DURATION_PERCENTILES = 'duration_percentiles',
   DURATION_DISTRIBUTION = 'duration_distribution',
@@ -91,10 +83,6 @@ export function EAPChartsWidget({transactionName}: EAPChartsWidgetProps) {
   const [selectedWidget, setSelectedWidget] = useState<EAPWidgetType>(
     EAPWidgetType.DURATION_BREAKDOWN
   );
-  const location = useLocation();
-  const spanCategoryUrlParam = decodeScalar(
-    location.query?.[SpanIndexedField.SPAN_CATEGORY]
-  );
 
   const options = useMemo(() => {
     return Object.entries(WIDGET_OPTIONS).map(([key, value]) => ({
@@ -105,49 +93,9 @@ export function EAPChartsWidget({transactionName}: EAPChartsWidgetProps) {
 
   const {title, description} = getWidgetContents(selectedWidget);
 
-  const query = new MutableSearch('');
-  query.addFilterValue('transaction', transactionName);
-  if (spanCategoryUrlParam) {
-    query.addFilterValue('span.category', spanCategoryUrlParam);
-  }
-
-  const {
-    data: spanSeriesData,
-    isPending,
-    error,
-  } = useSpanIndexedSeries(
-    {
-      yAxis: [
-        'avg(span.duration)',
-        'p100(span.duration)',
-        'p99(span.duration)',
-        'p95(span.duration)',
-        'p90(span.duration)',
-        'p75(span.duration)',
-        'p50(span.duration)',
-      ],
-      search: query,
-      transformAliasToInputFormat: true,
-    },
-
-    'transaction-summary-charts-widget',
-    DiscoverDatasets.SPANS_EAP
-  );
-
-  const timeSeries: TimeSeries[] = [];
-  Object.entries(spanSeriesData).forEach(([key, value]) => {
-    timeSeries.push({
-      field: key,
-      meta: {
-        type: value?.meta?.fields?.[key] ?? null,
-        unit: value?.meta?.units?.[key] as DataUnit,
-      },
-      data:
-        value?.data.map(item => ({
-          timestamp: item.name.toString(),
-          value: item.value,
-        })) ?? [],
-    });
+  const {plottables, isPending, error} = useWidgetChartQuery({
+    selectedWidget,
+    transactionName,
   });
 
   return (
@@ -162,9 +110,7 @@ export function EAPChartsWidget({transactionName}: EAPChartsWidgetProps) {
         isPending || error ? (
           <TimeSeriesWidgetVisualization.LoadingPlaceholder />
         ) : (
-          <TimeSeriesWidgetVisualization
-            plottables={timeSeries.map(series => new Area(series))}
-          />
+          <TimeSeriesWidgetVisualization plottables={plottables} />
         )
       }
       Footer={
